добавил что надо по чек листу
